Moving get_remote_data inside /lib
