+++ app/javascript/gobierto_dashboards/webapp/lib/get_remote_data.js
